requirements.txt
geopandas>=0.14.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pandas>=2.0.0
dbt-postgres>=1.7.0

